[Feat/#362] BannerDatailViewModel 파일 분리
